+++ libs/spring/npc/src/main/java/fr/guiguilechat/jcelechat/libs/spring/npc/agent/AgentRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AgentRepository extends JpaRepository<Agent, Long> {
+public interface AgentRepository extends JpaRepository<Agent, Integer> {
 
 	@Query("""
 select
